assimilate foreign project
