+++ .idea/compiler.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="HaskellBuildOptions">
+    <ghcPath>C:\Program Files\Haskell Platform\8.4.3\bin\ghc.exe</ghcPath>
+    <cabalPath>C:\Program Files\Haskell Platform\8.4.3\lib\extralibs\bin\cabal.exe</cabalPath>
+    <stackPath>C:\Users\tuv\AppData\Roaming\local\bin\stack.exe</stackPath>
+  </component>
+</project>
