--- README.md
@@ -1,6 +1,0 @@
-# Material-Hotel-Management-System
- A hotel managent system with javafx . with tabs for adding a client , reservation , search and paying . With simple and beautiful material design using fxml with css , we will update the software soon with a database connected and fully functional system.
-
-### Note
-if any problems with the jfoenix missing library , you can download it from here [jfoenix download](https://github.com/Amine-Smahi/MaterialLogin.Fx/raw/master/lib/jfoenix.jar)
-
