--- nbproject/private/private.properties
@@ -1,6 +1,0 @@
-auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
-# No need to modify this property unless customizing JavaFX Ant task infrastructure
-endorsed.javafx.ant.classpath=.
-javafx.run.inbrowser=<Default System Browser>
-javafx.run.inbrowser.path=C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe
-user.properties.file=C:\\Users\\DELL\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
