+++ src/app/Main.java
@@ -1,0 +1,40 @@
+package app;
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+import javafx.stage.StageStyle;
+
+/**
+ *
+ * @author DELL
+ */
+public class Main extends Application {
+    
+    @Override
+    public void start(Stage stage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("../fxml/FXMLDocument.fxml"));
+        stage.initStyle(StageStyle.UTILITY);
+        stage.centerOnScreen();
+        stage.setResizable(false);
+        Scene scene = new Scene(root);
+        stage.setAlwaysOnTop(true);
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        launch(args);
+    }
+    
+}
