+++ src/fxml/FXMLDocument.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane id="AnchorPane" prefHeight="606.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="test.on.dark.FXMLDocumentController">
+<AnchorPane id="AnchorPane" prefHeight="606.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="domain.FXMLDocumentController">
    <children>
       <JFXTabPane layoutY="69.0" prefHeight="537.0" prefWidth="900.0">
          <tabs>
